src/requirements.txt
anyio==3.3.4
asgiref==3.4.1
attrs==21.2.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
contextvars==2.4
cryptography==35.0.0
fastapi==0.70.0
fastapi-utils==0.2.1
greenlet==1.1.2
h11==0.12.0
idna==3.3
immutables==0.16
importlib-metadata==4.8.2
iniconfig==1.1.1
packaging==21.2
passlib==1.7.4
peewee==3.14.8
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.5
python-dotenv==0.19.2
requests==2.26.0
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
toml==0.10.2
typing==3.7.4.3
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
zipp==3.6.0